Integrate bb_parser with main library (builds but as of yet untested)
With this diff, the bb_parser "library" has in theory been integrated with the main firmware file.

This should mean that inbound messages that conform to the expected format should be parsed and set in local variables when shadow updates happen, and proper responses with the full JSON representing the state of the device / local data should now be generated and sent back.